% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrosintesis.R
\name{agrosintesis}
\alias{agrosintesis}
\title{Sintetiza resultados estadisticos para multiples variables}
\usage{
agrosintesis(
  data,
  variables,
  estructura = NULL,
  verbose = FALSE,
  report = TRUE,
  color = TRUE,
  ...
)
}
\arguments{
\item{data}{data.frame con los datos experimentales.}

\item{variables}{vector de nombres de variables respuesta.}

\item{estructura}{formula en texto tipo "Factor1~Factor2" para clusters.}

\item{verbose}{Logical. If TRUE, prints progress messages.}

\item{report}{logical, imprime reporte ejecutivo en consola.}

\item{color}{logical, usa color en el reporte (si esta disponible).}

\item{...}{argumentos adicionales pasados a \code{agrobox()}.}
}
\value{
- Sin clusters: lista con \code{tabla} y \code{decision}
- Con clusters: lista de dichas listas por cluster
}
\description{
Aplica \code{agrobox()} a una o varias variables y consolida
las tablas resumen. Genera ademas una sintesis decisional
orientada a interpretacion rapida.
}
